+++ src/App.tsx
@@ -1,17 +1,19 @@
 /* eslint-disable max-len */
 /* eslint-disable jsx-a11y/control-has-associated-label */
-import React, { useEffect, useMemo, useRef, useState } from 'react';
+import React, { useEffect, useMemo, useState } from 'react';
 import { UserWarning } from './UserWarning';
 import * as todosServices from './api/todos';
 import { Todo } from './types/Todo';
 import { ErrorType } from './types/Error';
 import classNames from 'classnames';
 import { handleError } from './services/ErrorHandling';
+import { Header } from './components/Header';
+import { Footer } from './components/Footer';
+import { TodosList } from './components/TodosList';
 
 export const App: React.FC = () => {
   const [todosList, setTodosList] = useState<Todo[]>([]);
   const [tempTodo, setTempTodo] = useState<Todo | null>(null);
-  const [title, setTitle] = useState('');
 
   const [activeTodo, setActiveTodo] = useState<Todo[]>([]);
   const [errorType, setErrorType] = useState<ErrorType | null>(null);
@@ -20,15 +22,13 @@ export const App: React.FC = () => {
   const [active, setActive] = useState(false);
   const [completed, setCompleted] = useState(false);
 
-  const inputField = useRef<HTMLInputElement | null>(null);
-
   const loadTodos = async () => {
     try {
       const todosData = await todosServices.getTodos().then(data => data);
 
       setTodosList(todosData);
     } catch (error) {
-      handleError(setErrorType, 'loading');
+      handleError(setErrorType, { type: 'loading', time: Date.now() });
       throw error;
     }
   };
@@ -52,53 +52,6 @@ export const App: React.FC = () => {
       .length;
   }, [todosList]);
 
-  const handleTitle = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setTitle(event.target.value);
-  };
-
-  const onSuccessfulRequest = () => {
-    setTitle('');
-    setActiveTodo([]);
-  };
-
-  const addTodos = (event: React.FormEvent<HTMLFormElement>) => {
-    handleError(setErrorType, null);
-    event.preventDefault();
-
-    if (title.trim().length === 0) {
-      handleError(setErrorType, 'empty');
-
-      return;
-    }
-
-    const newTodo: Omit<Todo, 'id'> = {
-      userId: todosServices.USER_ID,
-      title: title.trim(),
-      completed: false,
-    };
-
-    const temp = { id: 0, ...newTodo };
-
-    const request = todosServices.addTodos(newTodo);
-
-    setTempTodo(temp);
-    setActiveTodo([temp]);
-
-    request
-      .then((createdTodo: Todo) => {
-        setTodosList(prev => [...prev, createdTodo]);
-        onSuccessfulRequest();
-      })
-      .catch((error: Error) => {
-        handleError(setErrorType, 'add');
-        setActiveTodo([]);
-        throw error;
-      })
-      .finally(() => {
-        setTempTodo(null);
-      });
-  };
-
   const deleteTodo = async (todoId: number) => {
     handleError(setErrorType, null);
     try {
@@ -110,7 +63,7 @@ export const App: React.FC = () => {
         );
       });
     } catch (error) {
-      handleError(setErrorType, 'delete');
+      handleError(setErrorType, { type: 'delete', time: Date.now() });
       throw error;
     } finally {
       setActiveTodo([]);
@@ -144,7 +97,7 @@ export const App: React.FC = () => {
         });
       })
       .catch((error: Error) => {
-        handleError(setErrorType, 'update');
+        handleError(setErrorType, { type: 'update', time: Date.now() });
         throw error;
       })
       .finally(() => {
@@ -168,59 +121,14 @@ export const App: React.FC = () => {
     try {
       await Promise.all([...promiseArray]);
     } catch (error) {
-      handleError(setErrorType, 'update');
+      handleError(setErrorType, { type: 'update', time: Date.now() });
       throw error;
-    }
-  };
-
-  const clearCompleted = async () => {
-    const completedTodos = [...todosList].filter(current => current.completed);
-
-    const promiseArray = completedTodos.map((todo: Todo) =>
-      deleteTodo(todo.id),
-    );
-
-    setActiveTodo([...completedTodos]);
-
-    try {
-      await Promise.all([...promiseArray]);
-    } catch (error) {
-      handleError(setErrorType, 'delete');
-      throw error;
-    }
-  };
-
-  const filterTodos = (event: React.MouseEvent<HTMLAnchorElement>) => {
-    const filterParam: string = event.currentTarget.textContent;
-
-    if (filterParam === 'All') {
-      setAll(true);
-      setActive(false);
-      setCompleted(false);
-    }
-
-    if (filterParam === 'Active') {
-      setAll(false);
-      setActive(true);
-      setCompleted(false);
-    }
-
-    if (filterParam === 'Completed') {
-      setAll(false);
-      setActive(false);
-      setCompleted(true);
     }
   };
 
   useEffect(() => {
     loadTodos();
   }, []);
-
-  useEffect(() => {
-    if (inputField.current) {
-      inputField.current.focus();
-    }
-  }, [todos, activeTodo]);
 
   if (!todosServices.USER_ID) {
     return <UserWarning />;
@@ -231,257 +139,41 @@ export const App: React.FC = () => {
       <h1 className="todoapp__title">todos</h1>
 
       <div className="todoapp__content">
-        <header className="todoapp__header">
-          {/* this button should have `active` class only if all todos are completed */}
-          {todosList.length > 0 && (
-            <button
-              type="button"
-              className={classNames('todoapp__toggle-all', {
-                active: todosCounter === 0,
-              })}
-              data-cy="ToggleAllButton"
-              onClick={completeAllTodos}
-            />
-          )}
+        <Header
+          todosList={todosList}
+          todosCounter={todosCounter}
+          activeTodo={activeTodo}
+          setTodosList={setTodosList}
+          completeAll={completeAllTodos}
+          setErrorType={setErrorType}
+          setTempTodo={setTempTodo}
+          setActiveTodo={setActiveTodo}
+        />
 
-          {/* Add a todo on form submit */}
-          <form onSubmit={addTodos}>
-            <input
-              ref={inputField}
-              disabled={activeTodo.length > 0}
-              data-cy="NewTodoField"
-              type="text"
-              className="todoapp__new-todo"
-              placeholder="What needs to be done?"
-              value={title}
-              onChange={handleTitle}
-            />
-          </form>
-        </header>
-
-        <section className="todoapp__main" data-cy="TodoList">
-          {todos.map((todo: Todo) => {
-            return (
-              <div
-                key={todo.id}
-                data-cy="Todo"
-                className={classNames('todo', { completed: todo.completed })}
-              >
-                <label className="todo__status-label">
-                  <input
-                    data-cy="TodoStatus"
-                    type="checkbox"
-                    className="todo__status"
-                    checked={todo.completed}
-                    onChange={() => completeTodo(todo)}
-                  />
-                </label>
-
-                <span data-cy="TodoTitle" className="todo__title">
-                  {todo.title}
-                </span>
-                <button
-                  type="button"
-                  className="todo__remove"
-                  data-cy="TodoDelete"
-                  onClick={() => {
-                    deleteTodo(todo.id);
-                    setActiveTodo([todo]);
-                  }}
-                >
-                  ×
-                </button>
-
-                <div
-                  data-cy="TodoLoader"
-                  className={classNames('modal overlay', {
-                    'is-active': activeTodo.includes(todo),
-                  })}
-                >
-                  <div className="modal-background has-background-white-ter" />
-                  <div className="loader" />
-                </div>
-              </div>
-            );
-          })}
-          {tempTodo && (
-            <div key={tempTodo.id} data-cy="Todo" className="todo">
-              <label className="todo__status-label">
-                <input
-                  data-cy="TodoStatus"
-                  type="checkbox"
-                  className="todo__status"
-                />
-              </label>
-
-              <span data-cy="TodoTitle" className="todo__title">
-                {tempTodo.title}
-              </span>
-              <button
-                type="button"
-                className="todo__remove"
-                data-cy="TodoDelete"
-              >
-                ×
-              </button>
-
-              <div data-cy="TodoLoader" className="modal overlay is-active">
-                <div className="modal-background has-background-white-ter" />
-                <div className="loader" />
-              </div>
-            </div>
-          )}
-          {/* This is a completed todo */}
-          {/* <div data-cy="Todo" className="todo completed"> */}
-          {/* <label className="todo__status-label">
-              <input
-                data-cy="TodoStatus"
-                type="checkbox"
-                className="todo__status"
-                checked
-              />
-            </label>
-
-            <span data-cy="TodoTitle" className="todo__title">
-              Completed Todo
-            </span> */}
-
-          {/* Remove button appears only on hover */}
-          {/* <button type="button" className="todo__remove" data-cy="TodoDelete">
-              ×
-            </button> */}
-
-          {/* overlay will cover the todo while it is being deleted or updated */}
-          {/* <div data-cy="TodoLoader" className="modal overlay">
-              <div className="modal-background has-background-white-ter" />
-              <div className="loader" />
-            </div> */}
-          {/* </div> */}
-
-          {/* This todo is an active todo */}
-          {/* <div data-cy="Todo" className="todo">
-            <label className="todo__status-label">
-              <input
-                data-cy="TodoStatus"
-                type="checkbox"
-                className="todo__status"
-              />
-            </label>
-
-            <span data-cy="TodoTitle" className="todo__title">
-              Not Completed Todo
-            </span>
-            <button type="button" className="todo__remove" data-cy="TodoDelete">
-              ×
-            </button>
-
-            <div data-cy="TodoLoader" className="modal overlay">
-              <div className="modal-background has-background-white-ter" />
-              <div className="loader" />
-            </div>
-          </div> */}
-
-          {/* This todo is being edited */}
-          {/* <div data-cy="Todo" className="todo"> */}
-          {/* <label className="todo__status-label">
-              <input
-                data-cy="TodoStatus"
-                type="checkbox"
-                className="todo__status"
-              />
-            </label> */}
-
-          {/* This form is shown instead of the title and remove button */}
-          {/* <form>
-              <input
-                data-cy="TodoTitleField"
-                type="text"
-                className="todo__title-field"
-                placeholder="Empty todo will be deleted"
-                value="Todo is being edited now"
-              />
-            </form>
-
-            <div data-cy="TodoLoader" className="modal overlay">
-              <div className="modal-background has-background-white-ter" />
-              <div className="loader" />
-            </div> */}
-          {/* </div> */}
-
-          {/* This todo is in loadind state */}
-          {/* <div data-cy="Todo" className="todo"> */}
-          {/* <label className="todo__status-label">
-              <input
-                data-cy="TodoStatus"
-                type="checkbox"
-                className="todo__status"
-              />
-            </label>
-
-            <span data-cy="TodoTitle" className="todo__title">
-              Todo is being saved now
-            </span>
-
-            <button type="button" className="todo__remove" data-cy="TodoDelete">
-              ×
-            </button> */}
-
-          {/* 'is-active' class puts this modal on top of the todo */}
-          {/* <div data-cy="TodoLoader" className="modal overlay is-active">
-              <div className="modal-background has-background-white-ter" />
-              <div className="loader" />
-            </div> */}
-          {/* </div> */}
-        </section>
+        <TodosList
+          todos={todos}
+          activeTodo={activeTodo}
+          tempTodo={tempTodo}
+          deleteTodo={deleteTodo}
+          completeTodo={completeTodo}
+          setActiveTodo={setActiveTodo}
+        />
 
         {/* Hide the footer if there are no todos */}
         {todosList.length > 0 && (
-          <footer className="todoapp__footer" data-cy="Footer">
-            <span className="todo-count" data-cy="TodosCounter">
-              {`${todosCounter} items left`}
-            </span>
-
-            {/* Active link should have the 'selected' class */}
-            <nav className="filter" data-cy="Filter">
-              <a
-                href="#/"
-                className={classNames('filter__link', { selected: all })}
-                data-cy="FilterLinkAll"
-                onClick={filterTodos}
-              >
-                All
-              </a>
-
-              <a
-                href="#/active"
-                className={classNames('filter__link', { selected: active })}
-                data-cy="FilterLinkActive"
-                onClick={filterTodos}
-              >
-                Active
-              </a>
-
-              <a
-                href="#/completed"
-                className={classNames('filter__link', { selected: completed })}
-                data-cy="FilterLinkCompleted"
-                onClick={filterTodos}
-              >
-                Completed
-              </a>
-            </nav>
-
-            {/* this button should be disabled if there are no completed todos */}
-            <button
-              disabled={todosCounter === todosList.length}
-              type="button"
-              className="todoapp__clear-completed"
-              data-cy="ClearCompletedButton"
-              onClick={clearCompleted}
-            >
-              Clear completed
-            </button>
-          </footer>
+          <Footer
+            todosList={todosList}
+            todosCounter={todosCounter}
+            all={all}
+            active={active}
+            completed={completed}
+            onDelete={deleteTodo}
+            setAll={setAll}
+            setActive={setActive}
+            setCompleted={setCompleted}
+            setActiveTodo={setActiveTodo}
+            setErrorType={setErrorType}
+          />
         )}
       </div>
 
@@ -501,15 +193,15 @@ export const App: React.FC = () => {
           onClick={() => setErrorType(null)}
         />
         {/* show only one message at a time */}
-        {errorType === 'loading' && 'Unable to load todos'}
+        {errorType?.type === 'loading' && 'Unable to load todos'}
         <br />
-        {errorType === 'empty' && 'Title should not be empty'}
+        {errorType?.type === 'empty' && 'Title should not be empty'}
         <br />
-        {errorType === 'add' && 'Unable to add a todo'}
+        {errorType?.type === 'add' && 'Unable to add a todo'}
         <br />
-        {errorType === 'delete' && 'Unable to delete a todo'}
+        {errorType?.type === 'delete' && 'Unable to delete a todo'}
         <br />
-        {errorType === 'update' && 'Unable to update a todo'}
+        {errorType?.type === 'update' && 'Unable to update a todo'}
       </div>
     </div>
   );
