+++ src/components/TodosList/index.ts
@@ -1,0 +1,1 @@
+export { TodosList } from './TodosList';
